removendo o doble das moedas

+++ lib/features/estudante/loja/product_list_items.dart
@@ -162,7 +162,7 @@ class _ProductListScreenState extends State<ProductListScreen> {
         }
 
         transaction.update(userRef, {
-          'coins': FieldValue.increment(-totalCost),
+          'coins': FieldValue.increment(-totalCost.toInt()),
         });
 
         for (final cartItem in _cart.values) {
